feat(segment): migrate to the new component model
